Store group expansion state in model
This breaks model/view separation but the alternative would
be to create a filtered tree view which may be even more
complicated. This change also speeds up filtering operations by
caching and recycling group and item objects removed by the filter.
By doing this it is not necessary to make sure the view has processed
the changes for each removal/addition.