Add --start-in-systray: "Start application minimized in the system tray"
The aim of this option is to provide a way to start the Signet desktop
client e.g. during login and not have the user bothered by the window
popping up.
When needed the user shall display the window by either pressing the button
on the device, left-clicking the systray icon or chosing "Open" from the
systray icon right-click menu.

FIXME:
Currently showMinimized() is called before hide() to be able to make
the window visible when user presses the button on the device.
This introduces a flicker in the task bar since the window is for a brief
moment shown there as minimized before it is hidden.
It would be better to use setVisible() or showNormal() where the button
press is handled and just hide() in SignetApplication::init.