Add a git version string in the About dialog
The git version string is the output of
`git describe --tags --always --dirty` with "g:" prepended
or empty if we're not in a git repository, i.e. when building
from a tarball.

The script client/gitversion.sh is being called during the
build process to write the git version string into a define in
client/gitversion.h which is then used in client/about.cpp.