Initial firefox pluging
For now the plugin just transmits the current URL to the client via
WebSocket. The client reponds by selecting the account with the closest
matching URL. This passive approach can increase efficiency by removing
the need to search for the entry while maintaining security by not giving
any access to Signet over the Websocket.